goal tab add no goals part

+++ uReadLiteracy/CountingLabel.swift
@@ -15,9 +15,9 @@ class CountingLabel:UILabel{
     private let maxValue:Float
     private var numberOfUpdates:Double = 100
     private var percent:Int!
-    private let loadingDuration:Int
+    private let loadingDuration:Double
     
-    init(percent:Int,loadingDuration:Int, frame:CGRect){
+    init(percent:Int,loadingDuration:Double, frame:CGRect){
         self.maxValue = Float(percent)
         self.loadingDuration = loadingDuration
         self.percent = percent
